ml/training/requirements.txt
pandas
scikit-learn
lightgbm
clickhouse-connect
joblib
python-dotenv